12th_Nov/client-sentiment-radar/requirements.txt
fastapi
uvicorn
pandas
transformers
torch
requests
beautifulsoup4
lxml
streamlit
openai
python-dotenv
langchain
langchain-openai
reportlab
matplotlib
schedule
smtplib
